Fix home screen hang on Int overflow of media item position

+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/data/model/server/ServerMediaItem.kt
@@ -33,7 +33,7 @@ data class ServerMediaItem(
     // Album
     @SerialName("version") val version: String? = null,
     // @SerialName("external_ids") val externalIds: List<List<String>>? = null,
-    @SerialName("position") val position: Int? = null,
+    @SerialName("position") val position: Long? = null,
     @SerialName("year") val year: Int? = null,
     @SerialName("artists") val artists: List<ServerMediaItem>? = null,
     // @SerialName("album_type") val albumType: AlbumType? = null,
